requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
librosa>=0.10.0
soundfile>=0.12.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
jupyter>=1.0.0
ipywidgets>=8.0.0
dacite>=1.8.0
pytest>=7.0.0
